requirements.txt
Django==1.10.2
djangorestframework==3.5.3
inflect==0.2.5
Pillow==3.4.2
pytz==2016.7
shortuuid==0.4.3
